+++ backend/src/main/java/perf/shop/domain/order/application/OrdersService.java
@@ -26,7 +26,7 @@ public class OrdersService {
             OrderLine orderLine = OrderLine.of(newOrder, orderLineRequest);
             newOrder.addOrderLine(orderLine);
         });
-        
+
         ordersRepository.save(newOrder);
     }
 }
