+++ backend/src/main/java/perf/shop/domain/order/dto/request/OrderCreateRequest.java
@@ -6,6 +6,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import perf.shop.domain.model.dto.request.ShippingInfoRequest;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
