+++ backend/src/test/java/perf/shop/domain/order/api/OrdersApiTest.java
@@ -17,13 +17,13 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import perf.shop.domain.model.dto.request.AddressRequest;
+import perf.shop.domain.model.dto.request.ReceiverRequest;
+import perf.shop.domain.model.dto.request.ShippingInfoRequest;
 import perf.shop.domain.order.application.OrdersService;
-import perf.shop.domain.order.dto.request.AddressRequest;
 import perf.shop.domain.order.dto.request.OrderCreateRequest;
 import perf.shop.domain.order.dto.request.OrderLineRequest;
 import perf.shop.domain.order.dto.request.OrdererRequest;
-import perf.shop.domain.order.dto.request.ReceiverRequest;
-import perf.shop.domain.order.dto.request.ShippingInfoRequest;
 import perf.shop.global.common.response.ResponseCode;
 import perf.shop.global.error.exception.ErrorCode;
 import perf.shop.mock.InjectMockUser;
